# docs/swagger.yaml
basePath: /
definitions:
...
  response.Response:
    properties:
      code:
        description: 提示代码
        type: integer
      data:
        description: 数据
      msg:
        description: 提示信息
        type: string
    type: object
host: localhost:8082
info:
  contact:
    name: API Support
    url: https://github.com/carrot-hu23/dst-admin-go
  description: 饥荒联机版服务器管理后台 API 文档
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: DST Admin Go API
  version: "1.7.0"
paths:
...
  /api/dst/map/gen:
    get:
      description: 生成地图
      parameters:
      - description: levelName
        in: query
        name: levelName
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
      summary: 生成地图
      tags:
      - dstMap
  /api/dst/map/has/walrusHut/plains:
    get:
      description: 检测地图中是否有walrusHutPlains
      parameters:
      - description: levelName
        in: query
        name: levelName
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
      summary: 检测地图中是否有walrusHutPlains
      tags:
      - dstMap
  /api/dst/map/image:
    get:
      description: 获取地图图片
      parameters:
      - description: levelName
        in: query
        name: levelName
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
      summary: 获取地图图片
      tags:
      - dstMap
  /api/dst/map/player/session/file:
    get:
      description: 获取玩家存档文件
      parameters:
      - description: levelName
        in: query
        name: levelName
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
      summary: 获取玩家存档文件
      tags:
      - dstMap
  /api/dst/map/session/file:
    get:
      description: 获取存档文件
      parameters:
      - description: levelName
        in: query
        name: levelName
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
      summary: 获取存档文件
      tags:
      - dstMap
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"